Added IWebClientService and WebClientService
Added
The GET method is used to retrieve data from a server. When a client (such as a web browser) sends a GET request to a server, the server returns the requested data to the client. GET requests are used to retrieve information such as HTML pages, images, or other data. GET requests are typically used to retrieve data without changing it on the server.

The POST method is used to submit data to a server. When a client (such as a web browser) sends a POST request to a server, it sends data along with the request. The server then processes the data and returns a response to the client. POST requests are typically used to send data to the server for processing, such as submitting a form or uploading a file.

The PUT method is an HTTP method used to update a resource on a server. When a client sends a PUT request to a server, it includes the updated information in the request body. The server then updates the resource identified by the URL with the new data.